캠페인 컴포넌트 분리 및 참조 버그 수정
