+++ views/schedule-page.twig
@@ -401,6 +401,7 @@
         var scheduleRecurrenceDeleteUrl = "{{ url_for("schedule.recurrence.delete.form", {id:':id'}) }}";
         var layoutPreviewUrl = "{{ url_for("layout.preview", {id: ':id'}) }}";
         var layoutDesignerUrl = "{{ url_for("layout.designer", {id: ':id'}) }}";
+        var layoutSearchURL = "{{ url_for('layout.search') }}";
         var campaignEditFormUrl = "{{ url_for("campaign.edit.form", {id: ':id'}) }}";
         var scheduleSearchUrl = "{{ url_for("schedule.search") }}";
